requirements.txt
streamlit
agno
python-dotenv
requests